requirements.txt
dataclasses-json==0.5.6
kafka-python==2.0.2
marshmallow==3.13.0
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
names==0.3.0
typing-extensions==3.10.0.2
typing-inspect==0.7.1
